tcp: use sacks for timer based recovery
Type: feature

Stop distinguishing between fast and timer based recovery. That is,
reuse sack scoreboard, if available, to retransmit only those bytes
needed.

Additional changaes to:
- limited transmit updates
- add sacked rxt count to scoreboard

STILL NEEDED: recognize reneging

+++ src/vnet/session/transport.c
@@ -578,7 +578,9 @@ spacer_max_burst (spacer_t * pacer, u64 norm_time_now)
   u64 n_periods = norm_time_now - pacer->last_update;
   u64 inc;
 
-  if (n_periods > 0 && (inc = n_periods * pacer->tokens_per_period) > 10)
+//  clib_warning ("nperiods %u tokens/p %.2f inc %lu", n_periods,
+//                pacer->tokens_per_period, (u64) ((f32) n_periods * pacer->tokens_per_period));
+  if (n_periods > 0 && (inc = (f32) n_periods * pacer->tokens_per_period) > 10)
     {
       pacer->last_update = norm_time_now;
       pacer->bucket = clib_min (pacer->bucket + inc, pacer->bytes_per_sec);
